+++ src/test/resources/sql-schema.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS transactions
 (
     `id`               BIGINT AUTO_INCREMENT PRIMARY KEY,
     `transaction_date` DATE,
-    `vendor`           varchar(255) NOT NULL,
-    `amount`           DOUBLE       NOT NULL,
+    `vendor`           varchar(255)    NOT NULL,
+    `amount`           DECIMAL(6,2) NOT NULL,
     `category`         varchar(255)
 );
